test: user route test complete
